refactor(services): extract web agent run state and mission control
Moves the run fixtures and their write path into a state module and the
8 mission control methods into a client over a new interface file.
performMissionControlAction declares `this: AgentService` so its calls to
cancelAgentRun and resumeAgentRun stay late-bound.

The loop, chat and session contexts still project run state; they reach
it through projectWebOwnerRun, findWebAgentRun, prependWebAgentRun and
isTerminalWebRunState rather than through the binding.

web-agent-client.ts 2580 -> 2373, agent-service.ts 392 -> 384.
