+++ src/services/web-agent-client.ts
@@ -34,9 +34,8 @@ import { readWebAppSettings } from "./web-settings-client";
 import { defaultSessionTitleFromPath } from "../lib/session-path";
 import { snapshotSeat } from "./seat-presentation";
 import type { ChatConfig, ChatMessage, ChatStreamEvent } from "../types/chat";
-import type { AgentRun, AgentRunEvent } from "../types/agent-run";
+import type { AgentRun } from "../types/agent-run";
 import type { AgentRunnerDescriptor, AgentRunnerSelection } from "../types/agent-runner";
-import type { MissionControlActionReceipt, MissionControlOverview, MissionControlQuery, MissionControlRunDetail, MissionControlRunSummary } from "../types/mission-control";
 import { webEvaluationClient } from "./web-evaluation-client";
 import type {
   ContinueLoopInput,
@@ -98,6 +97,20 @@ import { webSessionCategoryClient } from "./web-session-category-client";
 import { webExpertRoleClient, listWebExpertRoles } from "./web-expert-role-client";
 import { webAgentTerminalClient } from "./web-agent-terminal-client";
 import { webUsageStatisticsClient } from "./web-usage-statistics-client";
+import { webMissionControlClient } from "./web-mission-control-client";
+import {
+  findWebAgentRun,
+  isTerminalWebRunState,
+  prependWebAgentRun,
+  projectWebOwnerRun,
+  setWebAgentRunEvents,
+  updateWebAgentRun,
+} from "./web-agent-run-state";
+
+export {
+  resetWebMissionControlRunsForTest,
+  seedWebMissionControlRunsForTest,
+} from "./web-agent-run-state";
 import {
   inspectMockProject,
   joinSiblingPath,
@@ -488,8 +501,8 @@ function finishWebGeneration(sessionId: string, lifecycleState: Session["lifecyc
   const session = findWebSession(sessionId);
   const runId = session.activeExecutionRunId;
   if (runId) {
-    const run = webAgentRuns.find((candidate) => candidate.id === runId);
-    if (run && !terminalRunStates.has(run.state)) {
+    const run = findWebAgentRun(runId);
+    if (run && !isTerminalWebRunState(run.state)) {
       updateWebAgentRun(
         run.id,
         run.version,
@@ -963,169 +976,6 @@ export function setWebLoopPhaseDelayForTest(delayMs: number): void {
 }
 
 const webCodeReviewClient = createWebCodeReviewClient(webSessionWorkspaceClient);
-const WEB_RUN_TIME = "2026-08-16T00:00:00.000Z";
-let webAgentRuns: AgentRun[] = [{
-  id: "018f0f17-4d6a-7e20-b41d-66c5271a28d0",
-  owner: { ownerType: "web_demo", ownerId: "web-session-open" },
-  links: [{ linkType: "session", linkId: "web-session-open" }],
-  parentRunId: null,
-  state: "paused",
-  recoveryPolicy: "not_recoverable",
-  runner: {
-    kind: "local", targetId: "local", targetRevision: null, label: "Local", hostLabel: "This device",
-    recovery: "none", capabilityWitness: "web-demo-local", authorityWitness: "web-demo-local", recoveryReference: null,
-  },
-  retryCount: 1,
-  maxRetries: 2,
-  reasonCode: "web_demo_paused",
-  createdAt: WEB_RUN_TIME,
-  updatedAt: WEB_RUN_TIME,
-  version: 4,
-  lastWitness: "web-demo-pause",
-}, ...([
-  ["waiting_approval", "approval_required"], ["waiting_user", "user_question"],
-  ["retrying", "provider_backoff"], ["stuck", "runner_disconnected"],
-  ["failed", "runner_interrupted"], ["completed", null], ["running", null],
-] as const).map(([state, reasonCode], index): AgentRun => ({
-  id: `018f0f17-4d6a-7e20-b41d-66c5271a29${index}`,
-  owner: { ownerType: index === 5 ? "evaluation" : "agent", ownerId: `web-owner-${index}` },
-  links: [{ linkType: "session", linkId: `web-session-${index}` }, ...(index === 4 ? [{ linkType: "review", linkId: "web-review-1" }] : [])],
-  parentRunId: null, state, recoveryPolicy: "owner_reconciles", retryCount: state === "retrying" ? 1 : 0,
-  maxRetries: 2, reasonCode, createdAt: `2026-08-16T00:0${index + 1}:00.000Z`,
-  updatedAt: `2026-08-16T00:0${index + 1}:30.000Z`, version: 2, lastWitness: `web-${state}`,
-  runner: index === 3 || index === 4 || index === 6 ? {
-    kind: "ssh", targetId: "web-demo-ssh", targetRevision: 1, label: "Build host", hostLabel: "build.example.test",
-    recovery: "inspect_only", capabilityWitness: "web-demo-ssh", authorityWitness: "web-demo-ssh-v1", recoveryReference: null,
-  } : undefined,
-}))];
-const defaultWebAgentRuns = structuredClone(webAgentRuns);
-const webAgentRunEvents = new Map<string, AgentRunEvent[]>([[webAgentRuns[0].id, [{
-  sequence: 4,
-  state: "paused",
-  trigger: "pause",
-  timestamp: WEB_RUN_TIME,
-  reasonCode: "web_demo_paused",
-  witness: "web-demo-pause",
-}]]]);
-
-export function seedWebMissionControlRunsForTest(count: 100 | 1_000): void {
-  const states: AgentRun["state"][] = [
-    "running", "waiting_approval", "waiting_user", "retrying", "blocked", "failed", "completed",
-  ];
-  webAgentRuns = Array.from({ length: count }, (_, index): AgentRun => {
-    const state = states[index % states.length];
-    const timestamp = new Date(Date.parse(WEB_RUN_TIME) + index * 1_000).toISOString();
-    return {
-      id: `018f0f17-4d6a-7e20-b41d-${String(index).padStart(12, "0")}`,
-      owner: { ownerType: "agent", ownerId: `performance-agent-${index % 10}` },
-      links: [{ linkType: "session", linkId: `performance-session-${index}` }],
-      parentRunId: null,
-      state,
-      recoveryPolicy: "owner_reconciles",
-      retryCount: state === "retrying" ? 1 : 0,
-      maxRetries: 2,
-      reasonCode: ["blocked", "failed"].includes(state) ? `performance_${state}` : null,
-      createdAt: timestamp,
-      updatedAt: timestamp,
-      version: 1,
-      lastWitness: `performance-fixture:${index}`,
-    };
-  });
-  webAgentRunEvents.clear();
-}
-
-export function resetWebMissionControlRunsForTest(): void {
-  webAgentRuns = structuredClone(defaultWebAgentRuns);
-  webAgentRunEvents.clear();
-  webAgentRunEvents.set(webAgentRuns[0].id, [{
-    sequence: 4,
-    state: "paused",
-    trigger: "pause",
-    timestamp: WEB_RUN_TIME,
-    reasonCode: "web_demo_paused",
-    witness: "web-demo-pause",
-  }]);
-}
-
-function updateWebAgentRun(runId: string, version: number, state: AgentRun["state"]): AgentRun {
-  const current = webAgentRuns.find((run) => run.id === runId);
-  if (!current) throw new Error(`run not found: ${runId}`);
-  if (["completed", "failed", "cancelled"].includes(current.state)) return current;
-  if (current.version !== version) throw new Error("run version conflict");
-  const nextVersion = version + 1;
-  const updatedAt = `2026-08-16T00:00:${String(nextVersion).padStart(2, "0")}.000Z`;
-  const updated = { ...current, state, reasonCode: null, version: nextVersion, updatedAt };
-  webAgentRuns = webAgentRuns.map((run) => run.id === runId ? updated : run);
-  const events = webAgentRunEvents.get(runId) ?? [];
-  events.push({
-    sequence: nextVersion,
-    state,
-    trigger: state === "cancelled" ? "cancel_user" : "resume",
-    timestamp: updatedAt,
-    reasonCode: null,
-    witness: `web-${state}:${runId}:${version}`,
-  });
-  webAgentRunEvents.set(runId, events);
-  return updated;
-}
-
-function projectWebOwnerRun(ownerId: string, state: AgentRun["state"]): void {
-  const run = webAgentRuns.find((item) => item.owner.ownerId === ownerId);
-  if (run && run.state !== state && !["completed", "failed", "cancelled"].includes(run.state)) {
-    updateWebAgentRun(run.id, run.version, state);
-  }
-}
-
-const terminalRunStates = new Set<AgentRun["state"]>(["completed", "failed", "cancelled"]);
-const activeRunStates = new Set<AgentRun["state"]>(["created", "preparing", "running", "waiting_approval", "waiting_user", "paused", "retrying", "blocked", "stuck", "verifying"]);
-
-function webMissionSummary(run: AgentRun): MissionControlRunSummary {
-  const session = run.links.find((link) => link.linkType === "session");
-  const review = run.links.find((link) => link.linkType === "review");
-  const attention = run.state === "waiting_approval" ? "approval" : run.state === "waiting_user" ? "user"
-    : ["blocked", "stuck"].includes(run.state) ? "stuck" : run.state === "failed" ? "failed" : review ? "review" : null;
-  const actions: MissionControlRunSummary["actions"] = ["open"];
-  if (!terminalRunStates.has(run.state)) actions.push("cancel");
-  if (["paused", "blocked", "stuck"].includes(run.state)) actions.push("resume");
-  if (["failed", "stuck"].includes(run.state) && run.retryCount < run.maxRetries) actions.push("retry");
-  if (run.state === "waiting_approval") actions.push("approval");
-  if (review) actions.push("review");
-  if (["completed", "failed"].includes(run.state)) actions.push("verify");
-  return {
-    runId: run.id, version: run.version, ownerType: run.owner.ownerType, ownerId: run.owner.ownerId,
-    agentId: run.owner.ownerType === "agent" ? run.owner.ownerId : null, title: `Run ${run.owner.ownerId}`,
-    state: run.state, createdAt: run.createdAt, updatedAt: run.updatedAt,
-    endedAt: terminalRunStates.has(run.state) ? run.updatedAt : null, projectId: null, workspace: null,
-    phase: run.state, attention, reasonCode: run.reasonCode,
-    verification: run.state === "verifying" ? "running" : run.state === "completed" ? "passed" : run.state === "failed" ? "failed" : "unavailable",
-    tokens: null, cost: null, actions,
-    navigation: review ? { kind: "review", id: review.linkId, sessionId: session?.linkId } : session ? { kind: "session", id: session.linkId } : null,
-    runner: run.runner ?? null,
-  };
-}
-
-function webMissionOverview(query: MissionControlQuery): MissionControlOverview {
-  const limit = Math.max(1, Math.min(query.limit ?? 20, 50));
-  const offset = query.cursor ? Number(query.cursor) : 0;
-  if (!Number.isSafeInteger(offset) || offset < 0) throw new Error("invalid mission control cursor");
-  let runs = webAgentRuns.map(webMissionSummary).filter((run) =>
-    (!query.states?.length || query.states.includes(run.state))
-    && (!query.agentId || run.agentId === query.agentId)
-    && (!query.projectId || run.projectId === query.projectId)
-    && (!query.runner || run.runner?.kind === query.runner));
-  const priority = (run: MissionControlRunSummary) => run.attention ? 0 : 1;
-  runs = runs.sort((left, right) => query.sort === "oldest"
-    ? left.createdAt.localeCompare(right.createdAt)
-    : query.sort === "attention" ? priority(left) - priority(right) || right.createdAt.localeCompare(left.createdAt)
-      : right.createdAt.localeCompare(left.createdAt));
-  const page = (items: MissionControlRunSummary[]) => ({ items: items.slice(offset, offset + limit), nextCursor: offset + limit < items.length ? String(offset + limit) : null });
-  const count = (state: AgentRun["state"]) => webAgentRuns.filter((run) => run.state === state).length;
-  return {
-    counts: { running: count("running"), waitingApproval: count("waiting_approval"), waitingUser: count("waiting_user"), retrying: count("retrying"), blocked: count("blocked") + count("stuck"), failed: count("failed"), completedRecently: count("completed") },
-    attention: page(runs.filter((run) => run.attention)), active: page(runs.filter((run) => activeRunStates.has(run.state))),
-    recent: page(runs.filter((run) => terminalRunStates.has(run.state))),
-  };
-}
 
 function webRunnerDescriptors(sessionId: string, agentId: string): AgentRunnerDescriptor[] {
   const session = findWebSession(sessionId);
@@ -1229,64 +1079,7 @@ export const webAgentClient: AgentService = {
   ...webSessionWorkspaceClient,
   ...webCodeReviewClient,
   ...webLspClient,
-  async getAgentRun(runId) {
-    const run = webAgentRuns.find((item) => item.id === runId);
-    if (!run) throw new Error(`run not found: ${runId}`);
-    return run;
-  },
-  async listAgentRuns(offset = 0, limit = 50, filter) {
-    const bounded = Math.max(1, Math.min(limit, 100));
-    const filtered = webAgentRuns.filter((run) =>
-      (!filter?.ownerType || run.owner.ownerType === filter.ownerType)
-      && (!filter?.ownerId || run.owner.ownerId === filter.ownerId)
-      && (!filter?.parentRunId || run.parentRunId === filter.parentRunId)
-      && (!filter?.state || run.state === filter.state));
-    return { items: filtered.slice(offset, offset + bounded), offset, limit: bounded };
-  },
-  async listAgentRunEvents(runId, offset = 0, limit = 50) {
-    const bounded = Math.max(1, Math.min(limit, 100));
-    return (webAgentRunEvents.get(runId) ?? []).slice(offset, offset + bounded);
-  },
-  async cancelAgentRun(runId, version) {
-    const cancelled = updateWebAgentRun(runId, version, "cancelled");
-    for (const child of webAgentRuns.filter((run) => run.parentRunId === runId)) {
-      if (!["completed", "failed", "cancelled"].includes(child.state)) {
-        updateWebAgentRun(child.id, child.version, "cancelled");
-      }
-    }
-    return cancelled;
-  },
-  async resumeAgentRun(runId, version) {
-    const run = webAgentRuns.find((item) => item.id === runId);
-    if (!run || !["paused", "blocked", "stuck"].includes(run.state)) throw new Error("run cannot be resumed");
-    return updateWebAgentRun(runId, version, "running");
-  },
-  async getMissionControlOverview(query = {}) { return structuredClone(webMissionOverview(query)); },
-  async getMissionControlRun(runId): Promise<MissionControlRunDetail> {
-    const run = webAgentRuns.find((item) => item.id === runId);
-    if (!run) throw new Error(`run not found: ${runId}`);
-    const linked = new Set(run.links.map((link) => link.linkType));
-    const facets: MissionControlRunDetail["facets"] = ["overview", "plan", "timeline", "tools", "files", "review", "verification", "context", "usage", "logs"].map((facet) => ({
-      facet: facet as MissionControlRunDetail["facets"][number]["facet"],
-      state: facet === "overview" || facet === "timeline" || facet === "logs" || linked.has(facet) ? "available" : "unavailable",
-    }));
-    return structuredClone({ run: webMissionSummary(run), facets });
-  },
-  async performMissionControlAction(input): Promise<MissionControlActionReceipt> {
-    const current = webAgentRuns.find((run) => run.id === input.runId);
-    if (!current) throw new Error(`run not found: ${input.runId}`);
-    if (current.version !== input.version) throw new Error("run version conflict");
-    if (input.action === "cancel") return { run: webMissionSummary(await this.cancelAgentRun(input.runId, input.version)), operationId: null };
-    if (input.action === "resume") return { run: webMissionSummary(await this.resumeAgentRun(input.runId, input.version)), operationId: null };
-    if (input.action === "retry") {
-      if (!["failed", "stuck"].includes(current.state) || current.retryCount >= current.maxRetries) throw new Error("run cannot be retried");
-      const retried = { ...current, id: `${current.id}-retry-${current.retryCount + 1}`, state: "retrying" as const, retryCount: current.retryCount + 1, version: 1, updatedAt: "2026-08-16T00:10:00.000Z", parentRunId: current.id, lastWitness: `web-retry:${current.id}` };
-      webAgentRuns = [retried, ...webAgentRuns];
-      return { run: webMissionSummary(retried), operationId: `web-retry-operation-${retried.id}` };
-    }
-    if (input.action === "verify") return { run: webMissionSummary(current), operationId: `web-verification-${current.id}` };
-    throw new Error("mission control action is unsupported");
-  },
+  ...webMissionControlClient,
 
   async deleteApiAgent(agentId: string) {
     if (mockAgents.find((agent) => agent.id === agentId)?.agentOrigin === "builtin") {
@@ -1576,7 +1369,7 @@ export const webAgentClient: AgentService = {
     };
     loopRuns = [run, ...loopRuns];
     const canonicalId = `018f0f17-4d6a-7e20-b41d-66c5271a${String(nextLoopRunId).padStart(4, "0")}`;
-    webAgentRuns = [{
+    prependWebAgentRun({
       id: canonicalId,
       owner: { ownerType: "loop_run", ownerId: runId },
       links: [{ linkType: "loop_definition", linkId: definitionId }],
@@ -1590,8 +1383,8 @@ export const webAgentClient: AgentService = {
       updatedAt: timestamp,
       version: 2,
       lastWitness: `web-loop-prepare:${runId}`,
-    }, ...webAgentRuns];
-    webAgentRunEvents.set(canonicalId, []);
+    });
+    setWebAgentRunEvents(canonicalId, []);
     emitLoopEvent(run);
     scheduleWebLoopPhase(run);
     return { run: cloneLoopValue(run), operationId };
@@ -1996,7 +1789,7 @@ export const webAgentClient: AgentService = {
       0,
     ) + 1;
     const executionRunId = `web-run-${input.sessionId}-${Date.now()}`;
-    webAgentRuns = [{
+    prependWebAgentRun({
       id: executionRunId,
       owner: { ownerType: "agent_generation", ownerId: session.agentId },
       links: [{ linkType: "session", linkId: input.sessionId }],
@@ -2011,8 +1804,8 @@ export const webAgentClient: AgentService = {
       updatedAt: timestamp,
       version: 2,
       lastWitness: `web-simulated-runner-start:${selectedRunner.selection.kind}`,
-    }, ...webAgentRuns];
-    webAgentRunEvents.set(executionRunId, []);
+    });
+    setWebAgentRunEvents(executionRunId, []);
     const userMessage: ChatMessage = {
       id: createMessageId(),
       sessionId: input.sessionId,
